+++ src/constants/api/base.js
@@ -1,0 +1,3 @@
+const BASE_URL = import.meta.env.VITE_BASE_URL;
+
+export default BASE_URL;
