+++ src/constants/errors/consoleErrors.js
@@ -1,0 +1,7 @@
+const CONSOLE_ERRORS = {
+  FETCH_FAILED: '데이터 요청 중 에러 발생:',
+  INVALID_RESPONSE: '서버 응답이 잘못되었습니다:',
+  UNSUPPORTED_METHOD: '지원하지 않는 요청 방식입니다:',
+};
+
+export default CONSOLE_ERRORS;
