+++ src/constants/errors/index.js
@@ -1,0 +1,2 @@
+export { default as CONSOLE_ERRORS } from './consoleErrors';
+export { default as UI_ERRORS } from './uiErrors';
