+++ src/constants/errors/uiErrors.js
@@ -1,4 +1,4 @@
-export const UI_ERRORS = {
+const UI_ERRORS = {
   NETWORK: '네트워크 오류가 발생했습니다. 다시 시도해주세요.',
   SERVER: '서버에서 문제가 발생했습니다. 관리자에게 문의하세요.',
   UNAUTHORIZED: '인증이 필요합니다. 다시 로그인해주세요.',
@@ -7,8 +7,4 @@ export const UI_ERRORS = {
   UNKNOWN: '알 수 없는 오류가 발생했습니다.',
 };
 
-export const CONSOLE_ERRORS = {
-  FETCH_FAILED: '데이터 요청 중 에러 발생:',
-  INVALID_RESPONSE: '서버 응답이 잘못되었습니다:',
-  UNSUPPORTED_METHOD: '지원하지 않는 요청 방식입니다:',
-};
+export default UI_ERRORS;
