+++ src/constants/index.js
@@ -1,3 +1,4 @@
-export * from './api';
-export * from './errors';
-export { default as STATUS_CODES } from './statusCodes';
+// 이렇게 하지 않고 각각의 폴더에서 바로 가져오는게 나을것 같아요.상의 후 삭제할지 말지 결정하겠습니다.
+// export * from './api';
+// export * from './errors';
+// export { default as STATUS_CODES } from './statusCodes';
