+++ src/constants/statusCodes/index.js
@@ -1,0 +1,1 @@
+export { default as STATUS_CODES } from './statusCodes';
