+++ src/pages/error/index.js
@@ -1,2 +1,2 @@
-export { default as NotFoundPage } from './NotFoundPage';
-export { default as ServerErrorPage } from './ServerErrorPage';
+export * from './notFound';
+export * from './serverError';
