+++ src/pages/error/notFound/index.js
@@ -1,0 +1,1 @@
+export { default as NotFoundPage } from './NotFoundPage';
