+++ src/pages/error/serverError/index.js
@@ -1,0 +1,1 @@
+export { default as ServerErrorPage } from './ServerErrorPage';
